requirements.txt
Flask==3.1.0
gunicorn==23.0.0
openai==1.58.1
pdfplumber==0.11.4
python-dotenv==1.0.1
stripe==11.4.1
